+++ app/assets/javascripts/ckeditor/plugins/wordcount/lang/en.js
@@ -4,10 +4,14 @@ For licensing, see LICENSE.html or http://ckeditor.com/license
 */
 CKEDITOR.plugins.setLang('wordcount', 'en', {
     WordCount: 'Words:',
+    WordCountRemaining: 'Words remaining',
     CharCount: 'Characters:',
+    CharCountRemaining: 'Characters remaining',
     CharCountWithHTML: 'Characters (with HTML):',
+    CharCountWithHTMLRemaining: 'Characters (with HTML) remaining',
     Paragraphs: 'Paragraphs:',
-    pasteWarning: 'Content can not be pasted because it is above the allowed limit',
+    ParagraphsRemaining: 'Paragraphs remaining',
+    pasteWarning: 'Content cannot be pasted because it is above the allowed limit',
     Selected: 'Selected: ',
     title: 'Statistics'
 });