+++ app/assets/javascripts/ckeditor/plugins/wordcount/lang/it.js
@@ -7,8 +7,9 @@ CKEDITOR.plugins.setLang('wordcount', 'it', {
     WordCount: 'Parole:',
     CharCount: 'Caratteri:',
     CharCountWithHTML: 'Caratteri (HTML incluso):',
-    Paragraphs: 'Paragraphs:',
-    pasteWarning: 'Content can not be pasted because it is above the allowed limit',
-    Selected: 'Selected: ',
+    Paragraphs: 'Paragrafi:',
+    ParagraphsRemaining: 'Paragraphs remaining',
+    pasteWarning: 'Il contenuto non può essere incollato poiché supera il limite massimo di caratteri disponibili',
+    Selected: 'Selezionato: ',
     title: 'Statistiche'
 });