+++ app/assets/javascripts/ckeditor/plugins/wordcount/lang/pl.js
@@ -6,8 +6,9 @@ CKEDITOR.plugins.setLang('wordcount', 'pl', {
     WordCount: 'Słów:',
     CharCount: 'Znaków:',
     CharCountWithHTML: 'Znaków (wraz z kodem HTML):',
-    Paragraphs: 'Paragraphs:',
-    pasteWarning: 'Content can not be pasted because it is above the allowed limit',
-    Selected: 'Selected: ',
+    Paragraphs: 'Paragrafy:',
+    ParagraphsRemaining: 'Paragraphs remaining',
+    pasteWarning: 'Zawartość nie może zostać wklejona, ponieważ przekracza dozwolony limit',
+    Selected: 'Zaznaczono: ',
     title: 'Statystyka'
 });